[Core] Move all core headers OUT of src/core
This will make true library installs a lot easier (aka: extremely infinitely more possible and less of a pain) but the current install code to do so is a bit broken atm. Copying the version.h should also be a thing since it has a potentional to be used by headers?